Untangle dispatcher/runner boundary violations inside internal/cli
Prepare the clicore extraction by making shared (CORE) files depend only
on other shared files, with no behavior change:

- Generalize argumentError's toCLIError into a classifiableCLIError
  interface so launch timeout errors classify themselves on the
  dispatcher side instead of being special-cased in the shared classifier
- Move unsupported-OS message mapping out of the shared classifier;
  install, update, and uninstall wrap their platform-guard errors in a
  self-classifying type (bootstrap_platform_errors.go)
- Move shared pieces into CORE files: the unityServerNotRespondingError
  family (server_not_responding.go), writeJSON, the option-help data
  layer (tool_option_help.go), the completion request predicate, pause
  point flag names, internal skill tool-name discovery
  (skill_sources.go), and the Unity package manifest schema
  (skills_packages.go)
