+++ cli/internal/cli/global_args.go
@@ -92,3 +92,19 @@ func containsHelpRequest(args []string) bool {
 	}
 	return false
 }
+
+// shouldHandleCompletionRequest reports whether args target the native completion
+// command or its list-commands/list-options helper flags, so run.go can route them
+// before falling back to the project-connected Unity tool dispatch path.
+func shouldHandleCompletionRequest(args []string) bool {
+	if len(args) == 0 {
+		return false
+	}
+
+	switch args[0] {
+	case listCommandsFlag, listOptionsFlag, completionCommand:
+		return true
+	default:
+		return false
+	}
+}
